feat: turn init module slug into a subcommand of the apply command
